CompanyController 의 공고관리 NCOmpanyController로 이동 등록폼 (파일 첨부 필요) / 공고 수정 (근무시간 안나오는 문제. convert생성하면 화면에 값이 나오지만 저장할 때 해결해야 할 문제가 존재.) 여러개 항목 값을 불러왔을 때 삭제하고 수정완료하면 오류가 뜨는 문제 존재 디테일 화면 수정 및 업데이트 중. /
